Refactor runtime package to be a vite library
