+++ packages/runtime/public/package/__init__.py
@@ -1,0 +1,2 @@
+def say_hello():
+    print("Hello from package")
