+++ packages/runtime/src/main.ts
@@ -1,10 +1,1 @@
-export { TerminalElement } from "./elements/terminal";
-export { EditorElement } from "./elements/editor";
-export { CodeElement } from "./elements/code";
-export { ControlsElement } from "./elements/controls";
-export { RunElement } from "./elements/run";
-export { WASIElement } from "./elements/wasi";
-export { FileElement } from "./elements/file";
-export { RunnoProvider } from "./provider";
-export { headlessRunCode, headlessRunFS } from "./headless";
-export { stripWhitespace, fetchWASIFS, elementCodeContent } from "./helpers";
+import "../lib/main";
