+++ packages/runtime/typedoc.config.cjs
@@ -1,6 +1,6 @@
 /** @type {import('typedoc').TypeDocOptions} */
 module.exports = {
-  entryPoints: ["./src/main.ts"],
+  entryPoints: ["./lib/main.ts"],
   out: "docs",
   // titleLink: "/",
   sidebarLinks: {
