+++ packages/wasi/package.json
@@ -29,9 +29,8 @@
     "dev": "vite --port 5173",
     "build:docs": "typedoc --options typedoc.config.cjs",
     "build:tests": "cd programs && cargo wasi build -r && cd .. && cp programs/target/wasm32-wasi/release/*.wasi.wasm public/bin/tests/",
-    "build:types": "tsc",
-    "build:vite": "vite build",
-    "build": "npm run build:docs && npm run build:vite && npm run build:types",
+    "build:vite": "tsc --noEmit && vite build",
+    "build": "npm run build:docs && npm run build:vite",
     "test:server": "vite --port 5173",
     "test:cargo": "cargo install cargo-wasi || echo Please install cargo to run the tests",
     "test:mkdir": "mkdir -p public/bin/languages public/bin/tests",
